refactor(frontend): Rename Ethereum WalletConnect send components
