+++ src/frontend/src/eth/components/wallet-connect/EthWalletConnectSendReview.svelte
